+++ src/society_elections/models/__init__.py
@@ -1,0 +1,6 @@
+from .candidate import Candidate
+from .election import Election
+from .electionposition import ElectionPosition
+from .position import Position
+from .vote import Vote
+from .voter import Voter
